move safe and folder decryption to helper functions
